Add support for Zabbix API v7
