+++ Cargo.toml
@@ -12,7 +12,7 @@ keywords = ["zabbix", "api"]
 categories = ["api-bindings"]
 
 [features]
-default = ["v7", "v6"]
+default = ["v7"]
 v7 = []
 v6 = []
