+++ src/client/mod.rs
@@ -1,6 +1,10 @@
 pub mod client;
 pub mod post;
+pub mod request;
 pub mod response;
+
+#[cfg(feature = "v7")]
+pub mod v7;
 
 #[cfg(feature = "v6")]
 pub mod v6;
