+++ src/client/post.rs
@@ -7,16 +7,26 @@ use crate::error::ZabbixApiError;
 const CONTENT_TYPE_HEADER: &str = "Content-Type";
 const CONTENT_TYPE_JSON: &str = "application/json";
 
-pub fn send_post_request<T: Serialize>(client: &Client,
-                                       url: &str, request: T) -> Result<String, ZabbixApiError> {
+pub fn send_post_request<T: Serialize>(
+    client: &Client,
+    url: &str,
+    session: Option<&str>,
+    request: T,
+) -> Result<String, ZabbixApiError> {
     debug!("send post request to '{url}'");
 
     let request_body = serde_json::to_string(&request)?;
 
-    let response = client.post(url)
+    let mut request = client
+        .post(url)
         .body(request_body)
-        .header(CONTENT_TYPE_HEADER, CONTENT_TYPE_JSON)
-        .send()?;
+        .header(CONTENT_TYPE_HEADER, CONTENT_TYPE_JSON);
+
+    if let Some(session) = session {
+        request = request.bearer_auth(session);
+    }
+
+    let response = request.send()?;
 
     let response_status = response.status();
     let response_text = response.text()?;
@@ -27,7 +37,6 @@ pub fn send_post_request<T: Serialize>(client: &Client,
 
     if response_status == reqwest::StatusCode::OK {
         Ok(response_text)
-
     } else {
         error!("unexpected server response code {}", response_status);
         Err(ZabbixApiError::BadRequestError)
